requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.0.7
django-allauth==0.63.6
django-appconf==1.0.6
django-cors-headers==4.4.0
django-filter==24.2
django-staticfiles==1.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pycparser==2.22
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.19.1
setuptools==72.1.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
